Safari-609.1.11.1/LayoutTests/fast/events/input-events-text-alignment-expected.txt

To manually test this, change the text alignment and check the resulting debug messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


beforeinput (inputType = formatJustifyCenter)
input (inputType = formatJustifyCenter)
getComputedStyle(text).textAlign = 'center'
beforeinput (inputType = formatJustifyRight)
input (inputType = formatJustifyRight)
getComputedStyle(text).textAlign = 'right'
beforeinput (inputType = formatJustifyLeft)
input (inputType = formatJustifyLeft)
getComputedStyle(text).textAlign = 'left'
beforeinput (inputType = formatJustifyFull)
input (inputType = formatJustifyFull)
getComputedStyle(text).textAlign = 'justify'
beforeinput (inputType = formatJustifyCenter)
Prevented default
getComputedStyle(text).textAlign = 'justify'
beforeinput (inputType = formatJustifyRight)
Prevented default
getComputedStyle(text).textAlign = 'justify'
beforeinput (inputType = formatJustifyLeft)
Prevented default
getComputedStyle(text).textAlign = 'justify'
beforeinput (inputType = formatJustifyFull)
Prevented default
getComputedStyle(text).textAlign = 'justify'
PASS successfullyParsed is true

TEST COMPLETE
WebKit
